refactor(protocols): break up protocols into separate files
